Adding scaling to the Serializer and Deserializer so that values can be scaled by a given amount to better fit the given data type. For example, if I need to send a floating point value between 0 and 1 I can give it a scale factor of 100 and sent it as a Uint8. Loses some precision, but saves some space. I've done this with all the network classes; they will now send information with less precise data types to save space.

+++ src/server/NetworkAsteroid.js
@@ -13,7 +13,7 @@ NetworkAsteroid.serializableProperties = [
   {key: 'x', type: 'Float32'},
   {key: 'y', type: 'Float32'},
   {key: 'colorIndex', type: 'Uint8'},
-  {key: 'radius', type: 'Float32'}
+  {key: 'radius', type: 'Uint16'}
 ];
 
 module.exports = NetworkAsteroid;
